refactor(modules): split into 5 modules with clean dep direction
Lantern is now a 5-module Go workspace:

- pb/    (github.com/anaregdesign/lantern/pb)    — generated proto stubs (leaf)
- core/  (github.com/anaregdesign/lantern/core)  — domain primitives (leaf)
- sdks/go/ (.../sdks/go) — client SDK, depends on pb only
- server/  (.../server)  — gRPC server, depends on pb + core
- .       (.../lantern)  — umbrella for cli + tests/integration

The previous layout had server/ depending on sdks/go/ purely to import
generated proto types (which lived inside the SDK module). That made the
dep graph point the wrong way (server should not depend on its client).

Fix: move generated stubs from sdks/go/gen/ into a new top-level pb/
module so both server and sdks/go depend on the shared contract.
buf.gen.yaml writes into pb/; sdks/go/ and server/ both import
github.com/anaregdesign/lantern/pb/graph/v1.

Also: the one server/service test that needed the client SDK (full
middleware bufconn round-trip) moved to tests/integration/ so server/
no longer requires sdks/go at all.

Wire tool directive lives in server/go.mod; `make wire` runs from server/.
Tag scheme gains pb/vX.Y.Z.
