+++ core/go.mod
@@ -1,0 +1,8 @@
+module github.com/anaregdesign/lantern/core
+
+go 1.26
+
+require (
+	github.com/google/uuid v1.6.0
+	golang.org/x/sync v0.20.0
+)
