+++ pb/go.mod
@@ -1,0 +1,17 @@
+module github.com/anaregdesign/lantern/pb
+
+go 1.26
+
+require (
+	github.com/grpc-ecosystem/grpc-gateway/v2 v2.29.0
+	google.golang.org/genproto/googleapis/api v0.0.0-20260526163538-3dc84a4a5aaa
+	google.golang.org/grpc v1.81.1
+	google.golang.org/protobuf v1.36.11
+)
+
+require (
+	golang.org/x/net v0.51.0 // indirect
+	golang.org/x/sys v0.42.0 // indirect
+	golang.org/x/text v0.36.0 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20260523011958-0a33c5d7ca68 // indirect
+)
