+++ server/service/service_bufconn_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	cachegraph "github.com/anaregdesign/lantern/core/cache/graph"
-	pb "github.com/anaregdesign/lantern/sdks/go/gen/graph/v1"
+	pb "github.com/anaregdesign/lantern/pb/graph/v1"
 	"github.com/anaregdesign/lantern/server/service"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
